attempt to disable security plugin rather than uninstall
see https://discuss.opendistrocommunity.dev/t/errors-opening-anomaly-detection-plugin-for-dashboards-after-creation-via-api/7711

I had hoped that perhaps the issue lay in the security plugin being removed. this commit disables rather than uninstalls it, but it behaves the same way. I'm committing it so I don't lose the work, but I'm going to revert it immediately