cisagov/Malcolm#460, work in progress for roles, arkime stuff
until Arkime v6.0.0 is out, as per @awick and I's discussion in slack, at the moment not all of the Arkime permissions can be set on roles, so creating these doesn't really do us any good. For now, then, Arkime roles are going to be handled purely based on URI path in the NGINX stuff (nginx/lua/nginx_auth_helpers.lua). Once all of these permissions are settable at the role level in Arkime, we can uncomment this and revisit it.