Improve the speed of pruning files
The pruning of files is too slow to keep up with the creation of extracted
files by Zeek. This change skips the slow gigabyte usage check when the
file system percent usage is exceeded. This change ensures that the
file system is not filled to capacity due to slow pruning of files.